Upgraded AppDynamics Enterprise Console, Java Agent, and Machine Agent to latest versions.

+++ provisioners/scripts/common/install_appdynamics_java_agent.sh
@@ -27,12 +27,12 @@ appd_controller_root_password="${appd_controller_root_password:-welcome1}"
 set -x  # turn command display back ON.
 appd_java_agent_home="${appd_java_agent_home:-appagent}"
 appd_java_agent_user="${appd_java_agent_user:-vagrant}"
-appd_java_agent_release="${appd_java_agent_release:-4.5.5.24422}"
+appd_java_agent_release="${appd_java_agent_release:-4.5.6.24621}"
 
 # [OPTIONAL] appdynamics java agent config parameters [w/ defaults].
 appd_java_agent_config="${appd_java_agent_config:-false}"
-appd_java_agent_controller_host="${appd_java_agent_controller_host:-apm}"
-appd_java_agent_controller_port="${appd_java_agent_controller_port:-8090}"
+appd_controller_host="${appd_controller_host:-apm}"
+appd_controller_port="${appd_controller_port:-8090}"
 appd_java_agent_application_name="${appd_java_agent_application_name:-My App}"
 appd_java_agent_tier_name="${appd_java_agent_tier_name:-My App Web Tier}"
 appd_java_agent_node_name="${appd_java_agent_node_name:-Development}"
@@ -43,37 +43,40 @@ appd_java_agent_account_access_key="${appd_java_agent_account_access_key:-abcdef
 usage() {
   cat <<EOF
 Usage:
-  All inputs are defined by external environment variables.
-  Optional variables have reasonable defaults, but you may override as needed.
-  Script should be run with 'root' privilege.
+  Install AppDynamics Java Agent by AppDynamics.
+
+  NOTE: All inputs are defined by external environment variables.
+        Optional variables have reasonable defaults, but you may override as needed.
+        Script should be run with 'root' privilege.
 
   -------------------------------------
   Description of Environment Variables:
   -------------------------------------
-   # [MANDATORY] appdynamics account parameters.
+  [MANDATORY] appdynamics account parameters.
     [root]# export appd_username="name@example.com"                     # user name for downloading binaries.
     [root]# export appd_password="password"                             # user password.
 
-   # [OPTIONAL] appdynamics java agent install parameters [w/ defaults].
+  [OPTIONAL] appdynamics java agent install parameters [w/ defaults].
     [root]# export appd_home="/opt/appdynamics"                         # [optional] appd home (defaults to '/opt/appdynamics').
     [root]# export appd_controller_root_password="welcome1"             # [optional] controller root password (defaults to 'welcome1').
     [root]# export appd_java_agent_home="appagent"                      # [optional] java agent home (defaults to 'appagent').
     [root]# export appd_java_agent_user="vagrant"                       # [optional] java agent user (defaults to user 'vagrant').
-    [root]# export appd_java_agent_release="4.5.5.24422"                # [optional] java agent release (defaults to '4.5.5.24422').
+    [root]# export appd_java_agent_release="4.5.6.24621"                # [optional] java agent release (defaults to '4.5.6.24621').
 
-   # [OPTIONAL] appdynamics java agent config parameters [w/ defaults].
+  [OPTIONAL] appdynamics java agent config parameters [w/ defaults].
     [root]# export appd_java_agent_config="true"                        # [optional] configure appd java agent? [boolean] (defaults to 'false').
 
-   # NOTE: Setting 'appd_java_agent_config' to 'true' allows you to perform the Java Agent configuration
-   #       concurrently with the installation. To successfully connect to the Controller, you should override
-   #       the 'appd_java_agent_controller_host' and 'appd_java_agent_account_access_key' parameters
-   #       using valid entries for your environment.
-   #
-   #       In either case, you will need to validate the configuration before starting the Java Agent. The
-   #       configuration file can be found here: '<java_agent_home>/appagent/ver4.5.5.24422/conf/controller-info.xml'
+    NOTE: Setting 'appd_java_agent_config' to 'true' allows you to perform the Java Agent configuration
+          concurrently with the installation. When 'true', the following environment variables are used
+          for the configuration. To successfully connect to the Controller, you should override the
+          'appd_controller_host' and 'appd_controller_port' parameters using valid entries for your
+          environment.
 
-    [root]# export appd_java_agent_controller_host="apm"                # [optional] controller host (defaults to 'apm').
-    [root]# export appd_java_agent_controller_port="8090"               # [optional] controller port (defaults to '8090').
+          In either case, you will need to validate the configuration before starting the Java Agent. The
+          configuration file can be found here: '<java_agent_home>/appagent/ver4.5.6.24621/conf/controller-info.xml'
+
+    [root]# export appd_controller_host="apm"                           # [optional] controller host (defaults to 'apm').
+    [root]# export appd_controller_port="8090"                          # [optional] controller port (defaults to '8090').
     [root]# export appd_java_agent_application_name="My App"            # [optional] associate java agent with application (defaults to ''My App).
     [root]# export appd_java_agent_tier_name="My App Web Tier"          # [optional] associate java agent with tier (defaults to 'My App Web Tier').
     [root]# export appd_java_agent_node_name="Development"              # [optional] associate java agent with node (defaults to 'Development').
@@ -106,24 +109,6 @@ set -x  # turn command display back ON.
 # set appdynamics java agent installation variables. -----------------------------------------------
 appd_java_agent_folder="${appd_java_agent_home}-${appd_java_agent_release}"
 appd_java_agent_binary="AppServerAgent-${appd_java_agent_release}.zip"
-
-# retrieve account access key from controller rest api if server is running.
-controller_url="http://${appd_java_agent_controller_host}:${appd_java_agent_controller_port}/controller/rest/serverstatus"
-controller_status=$(curl --silent --connect-timeout 10 ${controller_url} | awk '/available/ {print $0}' | awk -F ">" '{print $2}' | awk -F "<" '{print $1}')
-
-# if server is available, retrieve access key.
-if [ "$controller_status" == "true" ]; then
-  # build account info url to retrieve access key.
-  access_key_path="api/accounts/accountinfo?accountname=${appd_java_agent_account_name}"
-  access_key_url="http://${appd_java_agent_controller_host}:${appd_java_agent_controller_port}/${access_key_path}"
-
-  # retrieve the account access key from the returned json string.
-  set +x    # temporarily turn command display OFF.
-  controller_credentials="--user root@system:${appd_controller_root_password}"
-  access_key=$(curl --silent ${controller_credentials} ${access_key_url} | awk 'match($0,"accessKey") {print substr($0,RSTART-1,51)}')
-  set -x    # turn command display back ON.
-  appd_java_agent_account_access_key=$(echo ${access_key} | awk -F '"' '/accessKey/ {print $4}')
-fi
 
 # create appdynamics java agent parent folder. -----------------------------------------------------
 mkdir -p ${appd_home}/${appd_java_agent_folder}
@@ -170,15 +155,43 @@ chown -R ${appd_java_agent_user}:${appd_java_agent_user} .
 
 # configure appdynamics java agent. ----------------------------------------------------------------
 if [ "$appd_java_agent_config" == "true" ]; then
+  # set appdynamics java agent configuration variables.
+  appd_agent_config_path="${appd_home}/${appd_java_agent_home}/ver${appd_java_agent_release}/conf"
   appd_agent_config_file="controller-info.xml"
-  cd ${appd_home}/${appd_java_agent_home}/ver${appd_java_agent_release}/conf
-  cp -p ${appd_agent_config_file} ${appd_agent_config_file}.orig
 
-  sed -i -e "s/<controller-host>/<controller-host>${appd_java_agent_controller_host}/g" ${appd_agent_config_file}
-  sed -i -e "s/<controller-port>/<controller-port>${appd_java_agent_controller_port}/g" ${appd_agent_config_file}
-  sed -i -e "s/<application-name>/<application-name>${appd_java_agent_application_name}/g" ${appd_agent_config_file}
-  sed -i -e "s/<tier-name>/<tier-name>${appd_java_agent_tier_name}/g" ${appd_agent_config_file}
-  sed -i -e "s/<node-name>/<node-name>${appd_java_agent_node_name}/g" ${appd_agent_config_file}
-  sed -i -e "s/<account-name>/<account-name>${appd_java_agent_account_name}/g" ${appd_agent_config_file}
-  sed -i -e "s/<account-access-key>/<account-access-key>${appd_java_agent_account_access_key}/g" ${appd_agent_config_file}
+  cd ${appd_agent_config_path}
+
+  # save a copy of the current file.
+  if [ -f "${appd_agent_config_file}.orig" ]; then
+    cp -p ${appd_agent_config_file} ${appd_agent_config_file}.${curdate}
+  else
+    cp -p ${appd_agent_config_file} ${appd_agent_config_file}.orig
+  fi
+
+  # retrieve account access key from controller rest api if server is running.
+  controller_url="http://${appd_controller_host}:${appd_controller_port}/controller/rest/serverstatus"
+  controller_status=$(curl --silent --connect-timeout 10 ${controller_url} | awk '/available/ {print $0}' | awk -F ">" '{print $2}' | awk -F "<" '{print $1}')
+
+  # if server is available, retrieve access key.
+  if [ "$controller_status" == "true" ]; then
+    # build account info url to retrieve access key.
+    access_key_path="api/accounts/accountinfo?accountname=${appd_java_agent_account_name}"
+    access_key_url="http://${appd_controller_host}:${appd_controller_port}/${access_key_path}"
+
+    # retrieve the account access key from the returned json string.
+    set +x    # temporarily turn command display OFF.
+    controller_credentials="--user root@system:${appd_controller_root_password}"
+    access_key_record=$(curl --silent ${controller_credentials} ${access_key_url} | awk 'match($0,"accessKey") {print substr($0,RSTART-1,length($0)-2)}')
+    set -x    # turn command display back ON.
+    appd_java_agent_account_access_key=$(echo ${access_key_record} | awk -F '"' '/accessKey/ {print $4}')
+  fi
+
+  # use the stream editor to substitute the new values.
+  sed -i -e "/^    <controller-host>/s/^.*$/    <controller-host>${appd_controller_host}<\/controller-host>/" ${appd_agent_config_file}
+  sed -i -e "/^    <controller-port>/s/^.*$/    <controller-port>${appd_controller_port}<\/controller-port>/" ${appd_agent_config_file}
+  sed -i -e "/^    <application-name>/s/^.*$/    <application-name>${appd_java_agent_application_name}<\/application-name>/" ${appd_agent_config_file}
+  sed -i -e "/^    <tier-name>/s/^.*$/    <tier-name>${appd_java_agent_tier_name}<\/tier-name>/" ${appd_agent_config_file}
+  sed -i -e "/^    <node-name>/s/^.*$/    <node-name>${appd_java_agent_node_name}<\/node-name>/" ${appd_agent_config_file}
+  sed -i -e "/^    <account-name>/s/^.*$/    <account-name>${appd_java_agent_account_name}<\/account-name>/" ${appd_agent_config_file}
+  sed -i -e "/^    <account-access-key>/s/^.*$/    <account-access-key>${appd_java_agent_account_access_key}<\/account-access-key>/" ${appd_agent_config_file}
 fi
